pyproject.toml
[tool.poetry]
name = "personal-assistant"
version = "0.1.0"
description = ""
authors = ["juanlamadrid20 <enlightendevio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sounddevice = "^0.4.6"
numpy = "^1.26.4"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
python-docx = "^1.1.0"
torch = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
